LPD-76626 Move logic from GroupModelListener.java to CMSObjectEntryFolderDepotEntryLocalServiceWrapper.java to been able to filter y depot type
